adds NAUs Access (#388)
* fixes server loading issue (#378) (#379)

* changes in navigation drawer (#380)

* changes in navigation drawer

* implements full height navigation drawer

* fixes offline navigation issue (#377)

* Offline feature improvements (#381)

* handles error in downloading for offline files

* separate notifications for each complete offline file download

* Revert "Offline feature improvements (#381)" (#383)

This reverts commit c545ac4006d40e5ebf9a6a6c60a76190f75295be.

* adds support for recent files  (#374)

* adds recent files table

* adds recent files activity and its dependencies

* adds recent file actions

* adds upload support for all types of file (#385)

* Offline improvement (#384)

* handles error in downloading for offline files

* separate notifications for each complete offline file download

* changes in name and icons

* adds progress bar in offline activity

* fixes server name issue

* adds audio navigation (#387)

* adds audio navigation

* adds audio list

* improvements in cast (#386)

* adds local server login for secondary user

* adds HDA cache

* handling responses

* restoring session apis and fixing auto upload setting

* handles unauthorize response

* changes in TV UI for pin access

+++ src/main/java/org/amahi/anywhere/fragment/ServerFilesFragment.java
@@ -70,6 +70,7 @@ import org.amahi.anywhere.adapter.ServerFilesMetadataAdapter;
 import org.amahi.anywhere.bus.BusProvider;
 import org.amahi.anywhere.bus.FileOpeningEvent;
 import org.amahi.anywhere.bus.FileOptionClickEvent;
+import org.amahi.anywhere.bus.OfflineCanceledEvent;
 import org.amahi.anywhere.bus.OfflineFileDeleteEvent;
 import org.amahi.anywhere.bus.ServerFileDeleteEvent;
 import org.amahi.anywhere.bus.ServerFileDownloadingEvent;
@@ -83,6 +84,7 @@ import org.amahi.anywhere.server.client.ServerClient;
 import org.amahi.anywhere.server.model.ServerFile;
 import org.amahi.anywhere.server.model.ServerShare;
 import org.amahi.anywhere.util.Android;
+import org.amahi.anywhere.util.Downloader;
 import org.amahi.anywhere.util.Fragments;
 import org.amahi.anywhere.util.Intents;
 import org.amahi.anywhere.util.Mimes;
@@ -325,10 +327,6 @@ public class ServerFilesFragment extends Fragment implements
         BusProvider.getBus().post(new ServerFileSharingEvent(getShare(), getCheckedFile()));
     }
 
-    private ServerFile getOfflineServerFile(OfflineFile offlineFile) {
-        return new ServerFile(offlineFile.getName(), offlineFile.getTimeStamp(), offlineFile.getMime());
-    }
-
     private void deleteFile() {
         if (!isOfflineFragment()) {
             deleteFilePosition = getListAdapter().getSelectedPosition();
@@ -357,23 +355,26 @@ public class ServerFilesFragment extends Fragment implements
     }
 
     private void deleteFileFromOfflineStorage() {
-        OfflineFile offlineFile = mOfflineFileRepo.getOfflineFile(getShare().getName(), getCheckedFile().getPath(), getCheckedFile().getName());
-        if (offlineFile.getState() == OfflineFile.DOWNLOADING) {
-            stopDownloading(offlineFile.getDownloadId());
+        OfflineFile offlineFile = mOfflineFileRepo.getOfflineFile(getCheckedFile().getName(), getCheckedFile().getModificationTime().getTime());
+        if (offlineFile != null) {
+            if (offlineFile.getState() == OfflineFile.DOWNLOADING) {
+                stopDownloading(offlineFile.getDownloadId());
+            }
+            File file = new File(getContext().getFilesDir(), Downloader.OFFLINE_PATH + "/" + getCheckedFile().getName());
+            if (file.exists()) {
+                file.delete();
+            }
+            mOfflineFileRepo.delete(offlineFile);
+            Snackbar.make(getRecyclerView(), R.string.message_offline_file_deleted, Snackbar.LENGTH_SHORT)
+                .show();
         }
-        File file = new File(getContext().getFilesDir(), getCheckedFile().getName());
-        if (file.exists()) {
-            file.delete();
-        }
-        mOfflineFileRepo.delete(offlineFile);
-        Snackbar.make(getRecyclerView(), R.string.message_offline_file_deleted, Snackbar.LENGTH_SHORT)
-            .show();
     }
 
-    private void stopDownloading(long downloadID) {
+    private void stopDownloading(long downloadId) {
         DownloadManager dm = (DownloadManager) getContext().getSystemService(Context.DOWNLOAD_SERVICE);
         if (dm != null) {
-            dm.remove(downloadID);
+            dm.remove(downloadId);
+            BusProvider.getBus().post(new OfflineCanceledEvent(downloadId));
         }
     }
 
@@ -404,6 +405,7 @@ public class ServerFilesFragment extends Fragment implements
                 getFilesMetadataAdapter().removeFile(deleteFilePosition);
                 getFilesMetadataAdapter().setSelectedPosition(RecyclerView.NO_POSITION);
             }
+            showFilesContent();
         } else {
             Toast.makeText(getContext(), R.string.message_delete_file_error, Toast.LENGTH_SHORT).show();
         }
@@ -495,6 +497,7 @@ public class ServerFilesFragment extends Fragment implements
             }
         } else {
             getListAdapter().setAdapterMode(FilesFilterAdapter.AdapterMode.OFFLINE);
+            getFilesAdapter().setUpDownloadReceiver();
             showOfflineFiles();
         }
     }
@@ -937,7 +940,7 @@ public class ServerFilesFragment extends Fragment implements
     public void onMoreOptionClick(View view, int position) {
         setItemSelected(position);
         if (getListAdapter().getAdapterMode() != FilesFilterAdapter.AdapterMode.OFFLINE) {
-            Fragments.Builder.buildFileOptionsDialogFragment(getContext(), getCheckedFile())
+            Fragments.Builder.buildFileOptionsDialogFragment(getCheckedFile(), getShare())
                 .show(getChildFragmentManager(), "file_options_dialog");
         } else {
             Fragments.Builder.buildOfflineFileOptionsDialogFragment()
