+++ src/main/java/org/amahi/anywhere/service/DownloadService.java
@@ -1,5 +1,6 @@
 package org.amahi.anywhere.service;
 
+import android.app.DownloadManager;
 import android.app.Notification;
 import android.app.NotificationManager;
 import android.app.Service;
@@ -9,7 +10,9 @@ import android.net.ConnectivityManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.IBinder;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.v4.content.LocalBroadcastManager;
 import android.util.Log;
 
 import com.squareup.otto.Subscribe;
@@ -18,8 +21,11 @@ import org.amahi.anywhere.AmahiApplication;
 import org.amahi.anywhere.R;
 import org.amahi.anywhere.bus.BusProvider;
 import org.amahi.anywhere.bus.FileMovedEvent;
+import org.amahi.anywhere.bus.OfflineCanceledEvent;
 import org.amahi.anywhere.db.entities.OfflineFile;
+import org.amahi.anywhere.db.entities.RecentFile;
 import org.amahi.anywhere.db.repositories.OfflineFileRepository;
+import org.amahi.anywhere.db.repositories.RecentFileRepository;
 import org.amahi.anywhere.job.NetConnectivityJob;
 import org.amahi.anywhere.server.client.ServerClient;
 import org.amahi.anywhere.server.model.ServerFile;
@@ -30,12 +36,16 @@ import org.amahi.anywhere.util.Intents;
 import org.amahi.anywhere.util.NetworkUtils;
 
 import java.io.File;
+import java.util.List;
 
 import javax.inject.Inject;
 
 import static org.amahi.anywhere.util.Downloader.OFFLINE_PATH;
 
 public class DownloadService extends Service implements Downloader.DownloadCallbacks {
+
+    private static final int NOTIFICATION_OFFLINE_ID = 101;
+    private static final int NOTIFICATION_ERROR_ID = 101;
 
     @Inject
     ServerClient serverClient;
@@ -113,7 +123,7 @@ public class DownloadService extends Service implements Downloader.DownloadCallb
             scheduleNetworkConnectivityJob();
         }
 
-        return START_NOT_STICKY;
+        return START_STICKY;
     }
 
     private void startNextDownload() {
@@ -121,10 +131,12 @@ public class DownloadService extends Service implements Downloader.DownloadCallb
         if (file != null) {
             long downloadId = downloader.startDownloadingForOfflineMode(Uri.parse(file.getFileUri()), file.getName());
             saveOfflineFileDownloadId(file, downloadId);
+            notifyDownloadStart();
             Log.i("DownloadService", "Download id : " + file.getName() + downloadId);
             isDownloading = true;
         } else {
             stopDownloading();
+            stopForeground(true);
         }
     }
 
@@ -142,12 +154,12 @@ public class DownloadService extends Service implements Downloader.DownloadCallb
         notificationBuilder
             .setOngoing(true)
             .setSmallIcon(R.drawable.ic_app_logo)
-            .setContentTitle(getString(R.string.notification_download_title))
+            .setContentTitle(getString(R.string.notification_download_offline_title))
             .setContentText(getString(R.string.notification_upload_message, fileName))
             .setProgress(100, 0, false)
             .build();
         Notification notification = notificationBuilder.build();
-        startForeground((int) downloadId, notification);
+        startForeground(NOTIFICATION_OFFLINE_ID, notification);
         downloader.setDownloadCallbacks(this);
         downloader.resumeProgressCount(downloadId);
     }
@@ -161,14 +173,35 @@ public class DownloadService extends Service implements Downloader.DownloadCallb
     }
 
     private void saveFileInOfflineFileQueue(ServerFile serverFile, ServerShare serverShare) {
-        String downloadUri = serverClient.getFileUri(serverShare, serverFile).toString();
-        OfflineFile offlineFile = new OfflineFile(serverShare.getName(), serverFile.getPath(), serverFile.getName());
+
+        String downloadUri;
+        String shareName, path;
+        if (serverShare == null) {
+            // Request is from recent Files
+            downloadUri = getDownloadUriFrom(getRecentFile(serverFile.getUniqueKey()));
+            shareName = getRecentFile(serverFile.getUniqueKey()).getShareName();
+            path = getRecentFile(serverFile.getUniqueKey()).getPath();
+        } else {
+            downloadUri = serverClient.getFileUri(serverShare, serverFile).toString();
+            shareName = serverShare.getName();
+            path = serverFile.getPath();
+        }
+        OfflineFile offlineFile = new OfflineFile(shareName, path, serverFile.getName());
         offlineFile.setFileUri(downloadUri);
         offlineFile.setTimeStamp(serverFile.getModificationTime().getTime());
         offlineFile.setMime(serverFile.getMime());
         offlineFile.setState(OfflineFile.DOWNLOADING);
         offlineFile.setDownloadId(-1);
         offlineFileRepository.insert(offlineFile);
+    }
+
+    private RecentFile getRecentFile(String uniqueKey) {
+        RecentFileRepository repository = new RecentFileRepository(this);
+        return repository.getRecentFile(uniqueKey);
+    }
+
+    private String getDownloadUriFrom(RecentFile recentFile) {
+        return recentFile.getUri();
     }
 
     @Override
@@ -194,12 +227,17 @@ public class DownloadService extends Service implements Downloader.DownloadCallb
         notificationBuilder
             .setOngoing(true)
             .setSmallIcon(R.drawable.ic_app_logo)
-            .setContentTitle(getString(R.string.notification_download_title))
+            .setContentTitle(getString(R.string.notification_download_offline_title))
             .setContentText(getString(R.string.notification_upload_message, fileName))
             .setProgress(100, 0, false)
             .build();
         Notification notification = notificationBuilder.build();
-        startForeground(id, notification);
+        startForeground(NOTIFICATION_OFFLINE_ID, notification);
+    }
+
+    private void notifyDownloadStart() {
+        Intent intent = new Intent("DownloadStarted");
+        LocalBroadcastManager.getInstance(this).sendBroadcast(intent);
     }
 
     @Override
@@ -209,7 +247,14 @@ public class DownloadService extends Service implements Downloader.DownloadCallb
         notificationBuilder
             .setProgress(100, progress, false);
         Notification notification = notificationBuilder.build();
-        notificationManager.notify(id, notification);
+        notificationManager.notify(NOTIFICATION_OFFLINE_ID, notification);
+        updateOfflineActivityUI(progress);
+    }
+
+    private void updateOfflineActivityUI(int progress) {
+        Intent intent = new Intent("DownloadProgress");
+        intent.putExtra("progress", progress);
+        LocalBroadcastManager.getInstance(this).sendBroadcast(intent);
     }
 
     @Override
@@ -222,26 +267,38 @@ public class DownloadService extends Service implements Downloader.DownloadCallb
         }
     }
 
+    private void showDownloadedNotification(@NonNull OfflineFile offlineFile) {
+        NotificationManager notificationManager = (NotificationManager) getApplicationContext()
+            .getSystemService(Context.NOTIFICATION_SERVICE);
+
+        notificationBuilder
+            .setContentTitle(getString(R.string.notification_offline_download_complete))
+            .setContentText(getString(R.string.notification_upload_message, offlineFile.getName()))
+            .setOngoing(false);
+
+        Notification notification = notificationBuilder.build();
+        notificationManager.cancel(NOTIFICATION_OFFLINE_ID);
+        notificationManager.notify((int) offlineFile.getTimeStamp(), notification);
+    }
+
+    private void notifyDownloadFinish(boolean isSuccess) {
+        Intent intent = new Intent("DownloadFinished");
+        intent.putExtra("success", isSuccess);
+        LocalBroadcastManager.getInstance(this).sendBroadcast(intent);
+    }
+
     @Subscribe
     public void onFileMoved(FileMovedEvent event) {
         OfflineFile offlineFile = offlineFileRepository.getCurrentDownloadingFile();
         if (offlineFile != null) {
-
-            stopForeground(false);
-            NotificationManager notificationManager = (NotificationManager) getApplicationContext()
-                .getSystemService(Context.NOTIFICATION_SERVICE);
-
-            notificationBuilder
-                .setContentTitle(getString(R.string.notification_offline_download_complete))
-                .setOngoing(false)
-                .setProgress(100, 100, false);
-
-            Notification notification = notificationBuilder.build();
-            notificationManager.notify((int) offlineFile.getDownloadId(), notification);
-
             offlineFile.setState(OfflineFile.DOWNLOADED);
             offlineFile.setDownloadId(-1);
             offlineFileRepository.update(offlineFile);
+
+            stopForeground(true);
+
+            showDownloadedNotification(offlineFile);
+            notifyDownloadFinish(true);
             startNextDownload();
         }
     }
@@ -261,7 +318,39 @@ public class DownloadService extends Service implements Downloader.DownloadCallb
 
     @Override
     public void downloadError(long id) {
+        removeFromDownloader(id);
+        stopNotification();
+        showErrorNotification(id);
+        notifyDownloadFinish(false);
         removeOfflineFile(id);
+        startNextDownload();
+    }
+
+    private void removeFromDownloader(long id) {
+        DownloadManager dm = (DownloadManager) getSystemService(Context.DOWNLOAD_SERVICE);
+        if (dm != null) {
+            dm.remove(id);
+        }
+    }
+
+    private void showErrorNotification(long id) {
+        OfflineFile offlineFile = offlineFileRepository.getFileWithDownloadId(id);
+        if (offlineFile != null) {
+            NotificationManager notificationManager = (NotificationManager) getApplicationContext()
+                .getSystemService(Context.NOTIFICATION_SERVICE);
+
+            notificationBuilder
+                .setContentTitle(getString(R.string.notification_offline_download_error))
+                .setContentText(getString(R.string.notification_upload_message, offlineFile.getName()))
+                .setOngoing(false);
+
+            Notification notification = notificationBuilder.build();
+            notificationManager.notify((int) offlineFile.getTimeStamp(), notification);
+        }
+    }
+
+    private void stopNotification() {
+        stopForeground(true);
     }
 
     @Override
@@ -276,19 +365,32 @@ public class DownloadService extends Service implements Downloader.DownloadCallb
             .setProgress(100, progress, false);
 
         Notification notification = notificationBuilder.build();
-        notificationManager.notify((int) downloadId, notification);
+        notificationManager.notify(NOTIFICATION_OFFLINE_ID, notification);
 
         stopDownloading();
     }
 
     private void removeOfflineFile(long id) {
         OfflineFile offlineFile = offlineFileRepository.getFileWithDownloadId(id);
-        offlineFileRepository.delete(offlineFile);
+        if (offlineFile != null) {
+            offlineFileRepository.delete(offlineFile);
+        }
     }
 
     private void scheduleNetworkConnectivityJob() {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
             NetConnectivityJob.scheduleJob(this);
+        }
+    }
+
+    @Subscribe
+    public void onDownloadCanceled(OfflineCanceledEvent event) {
+        stopForeground(true);
+        List<OfflineFile> offlineFiles = offlineFileRepository.getAllOfflineFiles();
+        if (offlineFiles.isEmpty()) {
+            stopDownloading();
+        } else if (event.getDownloadId() != -1) {
+            startNextDownload();
         }
     }
 }
